Homework Logging and Configure in WebApp:
AvatarService:

INFO logs at the beginning of each method with the name of the method and parameters.

ERROR logs before throwing exceptions with additional error information.

WARNING logs for potential problems:

* Empty files or files without an extension
* Inability to read the image
* Incorrect file names

DEBUG logs for detailed information:

* File extension processing
* Creating directories
* Preview generation
* Normalization of file names
* Search/creating avatars

FacultyService:

INFO logs at the beginning of each method with the name of the method and parameters

ERROR logs before throwing the FacultyNotFoundException exception

DEBUG logs for detailed information:

* Faculty search results
* Number of faculties/students found
* The DTO mapping process
* Saving to the database
* WARNING logs (implicitly) via ERROR for cases when faculty is not found

StudentService:

INFO logs at the beginning of each method with the name of the method and parameters

ERROR logs before throwing the exception StudentNotFoundException

DEBUG logs for detailed information:

* Student Search Results
* The number of students found
* The DTO mapping process
* Saving to the database
* Statistical data (number, average age)
* Information about the faculties of the students

UniversityManagementService:

INFO logs at the beginning of each method with the name of the method and parameters

ERROR logs before throwing the exceptions FacultyNotFoundException and StudentNotFoundException

WARNING logs for cases when a student does not have a faculty

DEBUG logs for detailed information:

* Search results for faculties and students
* The number of students found for the faculty
* The process of removing faculty associations from students
* Information about found entities

Added 3 different application.properties files for different profiles.